requirements.txt
tensorflow>=2.13
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
opencv-python>=4.8
matplotlib>=3.7
tqdm>=4.65
PyYAML>=6.0
